# Use an efficient representation for merged components of operations