Currently [`merge_operations`](https://github.com/quantumlib/Cirq/blob/a89f5b9f0875dcae419efc13790af078075fcb6b/cirq-core/cirq/transformers/transformer_primitives.py#L357) represents merged components as a `CircuitOperation`. This means in a merge of `n` operations, `n-1` `CircuitOperations` are created, with a complexity of `O(n^2)`.
I use a disjont-set data structure to reduce the complexity to `O(n)` for `merge_k_qubit_unitaries_to_circuit_op`. `merge_operations` itself can't be improved because it uses a `merge_func` that requires creating a `CircuitOperation` at every step.