@codrut3 2025-07-13
I didn't change the algorithm in merge operations, but switched from CircuitOperation to Component for intermediate merges.
I wrote an explanation for the algorithm [here](https://docs.google.com/document/d/179nk8NMo-VtUgExSm73oU2oLLp7oGfUuVfZ1-RWiYi8/edit?usp=sharing).

I found this inefficiency while investigating why `merge_single_qubit_gates_to_phased_x_and_z` in the Shannon decomposition is slow. Right now this method is not called anymore by the Shannon decomposition, I replaced it with a custom version that is faster.

Still this change improves transpilation. On my laptop, I measured an average improvement of 5,09% from 86.89 seconds to 82,46 seconds for 7 qubits.

I also measured the improvement, if `merge_single_qubit_gates_to_phased_x_and_z` would still be used in the Shannon decomposition. In this case it is 20.7% from 40.88 seconds to 32.40 seconds for 8 qubits.

This is the script I used to measure the transpiler improvement:
```python
"""Tests performance of quantum Shannon decomposition.
"""

import sys
import time

from scipy.stats import unitary_group

import cirq


if not sys.warnoptions:
    import warnings
    warnings.simplefilter("ignore")


def main():
    measurements = []
    n_qubits = 7
    for i in range(10):
        print(f'Iteration {i}')
        U = unitary_group.rvs(2**n_qubits)
        qubits = [cirq.NamedQubit(f'q{i}') for i in range(n_qubits)]
        start_time = time.time()
        circuit = cirq.Circuit(cirq.quantum_shannon_decomposition(qubits, U))
        elapsed_time = time.time() - start_time
        print(f'Time taken: {elapsed_time:.2f} seconds')

        # Run cirq transpiler
        start_time = time.time()
        circuit = cirq.optimize_for_target_gateset(circuit, gateset=cirq.CZTargetGateset(preserve_moment_structure=False), max_num_passes=None)
        elapsed_time = time.time() - start_time
        measurements.append(elapsed_time)
        print(f'Cirq depth for {n_qubits} qubits (with transpiler): {len(circuit)} ({elapsed_time:.2f} sec)')

    print(f'Average time taken for transpiling: {sum(measurements) / len(measurements):.2f} seconds')


if __name__ == '__main__':
    main()
```