@codrut3 2025-08-10
> @codrut3 sorry about the late reply but I was wondering if instead of manually implementing union-find we can do the required the functionality using https://networkx.org/documentation/stable/_modules/networkx/utils/union_find.html
> 
> this will split the logic of merging moments into two parts where one of them where we need to maintain only one of them
> 
> ```python
>    def merge(self, a, b):
>           self._merge_moments(a, b)
>           self.union_find.union(a, b)
> ```
> 
> is it possible to do that or will it break the logic?

I think it works to use a separate data structure to do the union-find. This structure would have to be initialized by the user and passed as an argument to the component factory.

However, I would prefer to use [scipy.DisjointSet](https://docs.scipy.org/doc/scipy/reference/generated/scipy.cluster.hierarchy.DisjointSet.html) because it allows me to add new singletons on the fly, while https://networkx.org/documentation/stable/_modules/networkx/utils/union_find.html requires all elements to be known from the beginning.

If you are ok with adding a dependency on the scipy package, I will rewrite the code. Do I need to add the package version as a requirement somewhere?

Let me know how to proceed! @NoureldinYosri