@codrut3 2025-08-15
> @codrut3 that sounds good, Cirq already depends on scipy
> 
> https://github.com/quantumlib/Cirq/blob/fe0dc2187ca3269c178526e8ba41083fa1a467c9/cirq-core/requirements.txt#L10
> 
> sorry for the extra work

No worries! I changed the implementation to use `scipy.DisjointSet`.

I took this opportunity to rewrite the connected component classes. I moved the `merge()` and `find()` methods to the factory class, which I renamed as `ComponentSet`. This solved the issue that `can_merge` and `merge_func` were defined at the component level, and could potentially differ for the two components at merge time. Now they are defined at the level of `ComponentSet` which ensures all created components have the same callable (as well as are part of the same `scipy.DisjointSet`).