@pavoljuhas 2025-10-17
One last tweak in 7637ddea6e5d4eaa6bc44671adb32e84f5d8f3ca - typing_extensions is not in the cirq-core requirements anymore so I removed its import.